+++ foursight_core/fs_connection.py
@@ -46,18 +46,20 @@ class FSConnection(object):
         self.ff_env = fs_environ_info['ff_env']
         self.ff_es = fs_environ_info['es']
         self.ff_bucket = fs_environ_info['bucket']
+        self.redis = None
+        self.redis_url = None
         try:
             if 'redis' in fs_environ_info:
-                self.redis = RedisBase(create_redis_client(url=fs_environ_info['redis']))
+                self.redis_url = fs_environ_info['redis']
+                self.redis = RedisBase(create_redis_client(url=self.redis_url))
             elif 'REDIS_HOST' in os.environ:  # temporary patch in until env config is fully sorted - Will
-                self.redis = RedisBase(create_redis_client(url=os.environ['REDIS_HOST']))
-            else:
-                self.redis = None
+                self.redis_url = os.environ['REDIS_HOST']
+                self.redis = RedisBase(create_redis_client(url=self.redis_url))
         except redis.exceptions.ConnectionError:
             PRINT('Cannot connect to Redis')
             PRINT('This error is expected when deploying with remote (ElastiCache) Redis')
         PRINT(self.redis)
-        PRINT(os.environ.get('REDIS_HOST', 'no-redis-host'))
+        PRINT(self.redis_url)
         if not test:
             self.ff_s3 = s3Utils(env=self.ff_env)
             try:  # TODO: make this configurable from env variables?